add typecast to avoid (unlikely but possible) integer overflows.
this will silence a lot of static code analysis warnings

+++ src/KSPACE/pair_lj_cut_tip4p_long.cpp
@@ -604,7 +604,7 @@ void *PairLJCutTIP4PLong::extract(const char *str, int &dim)
 double PairLJCutTIP4PLong::memory_usage()
 {
   double bytes = maxeatom * sizeof(double);
-  bytes += maxvatom*6 * sizeof(double);
-  bytes += 2 * nmax * sizeof(double);
+  bytes += (double)maxvatom*6 * sizeof(double);
+  bytes += (double)2 * nmax * sizeof(double);
   return bytes;
 }
