+++ src/USER-REAXC/fix_qeq_reax.cpp
@@ -900,13 +900,13 @@ double FixQEqReax::memory_usage()
   double bytes;
 
   bytes = atom->nmax*nprev*2 * sizeof(double); // s_hist & t_hist
-  bytes += atom->nmax*11 * sizeof(double); // storage
-  bytes += n_cap*2 * sizeof(int); // matrix...
-  bytes += m_cap * sizeof(int);
-  bytes += m_cap * sizeof(double);
+  bytes += (double)atom->nmax*11 * sizeof(double); // storage
+  bytes += (double)n_cap*2 * sizeof(int); // matrix...
+  bytes += (double)m_cap * sizeof(int);
+  bytes += (double)m_cap * sizeof(double);
 
   if (dual_enabled)
-    bytes += atom->nmax*4 * sizeof(double); // double size for q, d, r, and p
+    bytes += (double)atom->nmax*4 * sizeof(double); // double size for q, d, r, and p
 
   return bytes;
 }
